site/apis/feedback/requirements.txt
flask>=2.0.2,<3.0.0
google-cloud-spanner>=3.15.1,<4.0.0
gunicorn>=20.1.0,<21.0.0
requests>=2.26.0,<3.0.0